+++ packages/app-admin/src/types.ts
@@ -7,15 +7,6 @@ type RenderParams = {
     content: React.ReactNode;
 };
 
-export type AdminWelcomeScreenWidgetPlugin = Plugin & {
-    type: "admin-welcome-screen-widget";
-    widget: {
-        title: string;
-        description: string;
-        cta: React.ReactNode;
-    };
-};
-
 export type AdminLayoutComponentPlugin = Plugin & {
     type: "admin-layout-component";
     render(params: RenderParams): React.ReactNode;