+++ packages/app-page-builder/src/admin/plugins/welcomeScreenWidget.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Link } from "@webiny/react-router";
 
-import { AdminWelcomeScreenWidgetPlugin } from "@webiny/app-admin/types";
+import { AdminWelcomeScreenWidgetPlugin } from "@webiny/app-plugin-admin-welcome-screen/types";
 import { ButtonSecondary } from "@webiny/ui/Button";
 import { css } from "emotion";